Use longer representation of stack traces in log messages
Previously, stack traces got truncated to the first 42 lines. Now
stacktraces up to a size of 55k are entirely displayed in Kibana.
If the size of a stacktrace exceeds 55k, the stacktrace will be
abbreviated in such a way that its first lines and its last lines
(which are considered the most important ones) are still logged.
Some lines in between are omitted in order to keep the overall
size of the stacktrace under 55k. The reason for this is that
larger log messages will not be processed successfully by the
logging service.

Change-Id: Idce5716bf4591bf235b852b5aa95ce854d68b66a