Initial draft to support tenant id
We've got several requests to support an additional (header) field
that would allow applications to attribute individual logs to a
"tenant".
This is an initial set of changes to support that where the HTTP
header field (name) is fixed ("tenantid"). As we have seen different
ways (names) to refer to the same concept it seems worthwhile to
investigate whether the mapping could be made configurable.