Add MDC fields to custom fields for logback
This change introduces a new configuration tag for the encoder:
`<customFieldMdcKeyName>`. An example configuration is

```
<encoder class="com.sap.hcp.cf.logback.encoder.JsonEncoder">
   	<customFieldMdcKeyName>custom-field</customFieldMdcKeyName>
   	<customFieldMdcKeyName>test-field</customFieldMdcKeyName>
</encoder>
```

During the marshalling of a log messages the MDC will be inspected for the
fields *custom-field* and *test-field*. If present, they will be written in the
nested *custom_fields* object of the output json. This feature allows for
special parsing of this nested data structure.