Migrate RequestLoggingFilter to new implementation
The RequestLoggingFilter is the originally provided filter of the module.
Goal of the refactoring was, to provide its functionality by several stand-alone filters.
To enable backwards-compatibility and ease of use, the original filter was
reimplemented as composite filter.

This can be used as an example for custom filter configurations and implementations.