Switch Custom Metrics Shipment
SAP BTP Application Logging has been changed to accept custom metrics within
the log message stream. For this the messages need to be in JSON format and
contain a field "type":"metrics" and an array of metrics in field "metrics". The prior
approach using an Http endpoint is deprecated for removal. This commit changes
cf-java-logging-support to provide custom metrics according to the new format by
writing the correct JSON to `System.out`.