Support for Http Proxy Headers (#119)
* Add logging of x-forwarded-* and x-custom-host http headers

Users can choose to log the following for http headers in the request logs:

* x-forwarded-for
* x-forwarded-host
* x-forwarded-proto
* x-custom-host

This requires the environment variable `LOG_SENSITIVE_CONNECTION_DATA` to be set to `true`.
If there are no headers, no fields will be added to the request logs. If there are headers, but
`LOG_SENSITIVE_CONNECTION_DATA` is set to `false`(default) the values will be logged as 'redacted'.

* Log SSL connection headers

HA-Proxy can add certain headers for terminated ssl connections [1].
These headers can be added to the request logs if environment variable
LOG_SSL_HEADERS is set to be true. Otherwise no fields will be added
to the request logs.

[1] https://www.haproxy.com/blog/ssl-client-certificate-information-in-http-headers-and-logs/

Co-authored-by: Christian Dinse <19688125+christiand93@users.noreply.github.com>