#135 Fix UOE when using Custom Fields (#136)
Log4j can provide different context data depending on its configuration.
In one case the provided context map is unmodifiable. In that case an
exception occurred, when a custom field was used as a message parameter.
This change guards against that issue, by defensively creating a modifiable
map of properties. Then the custom fields can be added safely.