Add Exception Type and Message to JSON Logs
Adds the two fields exception_type and exception_message to
the BaseFieldSuppliers for Log4j and Logback. This allows users
to override the provided values with custom FieldSuppliers. Basic
Unit Tests were added.

Performance implications should be minimal for events without
exceptions. If there are exceptions, the only longer check is for
the message to be non-blank.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>